+++ src/Vsix/Helpers/ProjectHelpers.cs
@@ -1,0 +1,137 @@
+﻿using EnvDTE;
+using EnvDTE80;
+using Microsoft.VisualStudio.Shell;
+using System;
+using System.IO;
+
+namespace EditorConfig
+{
+    public static class ProjectHelpers
+    {
+        static DTE2 _dte = Package.GetGlobalService(typeof(DTE)) as DTE2;
+
+        public static string GetRootFolder(this Project project)
+        {
+            if (project == null)
+                return null;
+
+            if (project.IsKind(ProjectKinds.vsProjectKindSolutionFolder))
+                return Path.GetDirectoryName(_dte.Solution.FullName);
+
+            if (string.IsNullOrEmpty(project.FullName))
+                return null;
+
+            string fullPath;
+
+            try
+            {
+                fullPath = project.Properties.Item("FullPath").Value as string;
+            }
+            catch (ArgumentException)
+            {
+                try
+                {
+                    // MFC projects don't have FullPath, and there seems to be no way to query existence
+                    fullPath = project.Properties.Item("ProjectDirectory").Value as string;
+                }
+                catch (ArgumentException)
+                {
+                    // Installer projects have a ProjectPath.
+                    fullPath = project.Properties.Item("ProjectPath").Value as string;
+                }
+            }
+
+            if (string.IsNullOrEmpty(fullPath))
+                return File.Exists(project.FullName) ? Path.GetDirectoryName(project.FullName) : null;
+
+            if (Directory.Exists(fullPath))
+                return fullPath;
+
+            if (File.Exists(fullPath))
+                return Path.GetDirectoryName(fullPath);
+
+            return null;
+        }
+
+        public static ProjectItem AddFileToProject(this Project project, string file, string itemType = null)
+        {
+            if (project.IsKind(ProjectTypes.ASPNET_5, ProjectTypes.SSDT))
+                return _dte.Solution.FindProjectItem(file);
+
+            var root = project.GetRootFolder();
+
+            if (string.IsNullOrEmpty(root) || !file.StartsWith(root, StringComparison.OrdinalIgnoreCase))
+                return null;
+
+            ProjectItem item = project.ProjectItems.AddFromFile(file);
+            item.SetItemType(itemType);
+            return item;
+        }
+
+        public static void SetItemType(this ProjectItem item, string itemType)
+        {
+            try
+            {
+                if (item == null || item.ContainingProject == null)
+                    return;
+
+                if (string.IsNullOrEmpty(itemType)
+                    || item.ContainingProject.IsKind(ProjectTypes.WEBSITE_PROJECT)
+                    || item.ContainingProject.IsKind(ProjectTypes.UNIVERSAL_APP))
+                    return;
+
+                item.Properties.Item("ItemType").Value = itemType;
+            }
+            catch (Exception ex)
+            {
+                System.Diagnostics.Debug.Write(ex);
+            }
+        }
+
+        public static bool IsKind(this Project project, params string[] kindGuids)
+        {
+            foreach (var guid in kindGuids)
+            {
+                if (project.Kind.Equals(guid, StringComparison.OrdinalIgnoreCase))
+                    return true;
+            }
+
+            return false;
+        }
+
+        public static string GetSelectedItemPath(out object selectedItem)
+        {
+            var items = (Array)_dte.ToolWindows.SolutionExplorer.SelectedItems;
+            selectedItem = null;
+
+            foreach (UIHierarchyItem selItem in items)
+            {
+                selectedItem = selItem.Object;
+
+                if (selItem.Object is ProjectItem item && item.Properties != null)
+                {
+                    return item.Properties.Item("FullPath").Value.ToString();
+                }
+                else if (selItem.Object is Project proj && proj.Kind != ProjectKinds.vsProjectKindSolutionFolder)
+                {
+                    return proj.GetRootFolder();
+                }
+                else if (selItem.Object is Solution sol && !string.IsNullOrEmpty(sol.FullName))
+                {
+                    return Path.GetDirectoryName(sol.FullName);
+                }
+            }
+
+            return null;
+        }
+    }
+
+    public static class ProjectTypes
+    {
+        public const string ASPNET_5 = "{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}";
+        public const string WEBSITE_PROJECT = "{E24C65DC-7377-472B-9ABA-BC803B73C61A}";
+        public const string UNIVERSAL_APP = "{262852C6-CD72-467D-83FE-5EEB1973A190}";
+        public const string NODE_JS = "{9092AA53-FB77-4645-B42D-1CCCA6BD08BD}";
+        public const string SSDT = "{00d1a9c2-b5f0-4af3-8072-f6c62b433612}";
+    }
+}